challenges/web/bruzz/Docker/web/src/requirements.txt
PyMySQL
Flask-MySQL
Flask-Session
